Fix build: LineItem.unitQty expects Int, not Long (#139)
Same fix as claude/ordermate-tickets-history-quantity-jdi5nn (cherry-picked
manually since this branch forked before that fix landed): a real build
caught LineItem.unitQty's setter expecting Int, not Long. Drop the
incorrect .toLong() conversion in both places the Quantity widget writes
it back.

+++ app/src/main/java/com/orderMate/activities/OverlayActivity.kt
@@ -128,7 +128,7 @@ class OverlayActivity : AppCompatActivity(), ILineItemUpdateListener {
                             allLineItems.forEach { lineItem ->
                                 if (lineItem?.item?.id == itemId) {
                                     lineItem.note = note
-                                    lineItem.unitQty = quantity.toLong()
+                                    lineItem.unitQty = quantity
                                 }
                             }
